Moved ProjectInfo creation from UploadBomMojo to ProjectAction
