+++ src/test/java/io/github/pmckeown/dependencytrack/upload/UploadBomMojoTest.java
@@ -1,16 +1,9 @@
 package io.github.pmckeown.dependencytrack.upload;
 
-import static io.github.pmckeown.dependencytrack.project.ProjectBuilder.aProject;
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
-
-import java.io.File;
-
+import io.github.pmckeown.dependencytrack.CommonConfig;
+import io.github.pmckeown.dependencytrack.metrics.MetricsAction;
+import io.github.pmckeown.dependencytrack.project.ProjectAction;
+import io.github.pmckeown.util.Logger;
 import org.apache.maven.project.MavenProject;
 import org.junit.Before;
 import org.junit.Test;
@@ -20,11 +13,16 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import io.github.pmckeown.dependencytrack.CommonConfig;
-import io.github.pmckeown.dependencytrack.metrics.MetricsAction;
-import io.github.pmckeown.dependencytrack.project.ProjectAction;
-import io.github.pmckeown.dependencytrack.project.ProjectInfo;
-import io.github.pmckeown.util.Logger;
+import java.io.File;
+
+import static io.github.pmckeown.dependencytrack.project.ProjectBuilder.aProject;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 
 @RunWith(MockitoJUnitRunner.class)
 public class UploadBomMojoTest {
@@ -87,26 +85,5 @@ public class UploadBomMojoTest {
         verifyNoInteractions(uploadBomAction);
         verifyNoInteractions(metricsAction);
         verifyNoInteractions(projectAction);
-    }
-
-    @Test
-    public void projectInfoCreationFromSbom() {
-        uploadBomMojo.setBomLocation(UploadBomMojoTest.class.getResource("bom.xml").getFile());
-        ProjectInfo info = uploadBomMojo.createProjectInfo().get();
-        assertThat(info.getGroup(), is(equalTo("io.github.pmckeown")));
-        assertThat(info.getDescription(), is(equalTo("Maven plugin to integrate with a Dependency Track server to submit dependency manifests and gather project metrics.")));
-        assertThat(info.getPurl(), is(equalTo("pkg:maven/io.github.pmckeown/dependency-track-maven-plugin@1.2.1-SNAPSHOT?type=maven-plugin")));
-        assertThat(info.getClassifier(), is(equalTo("LIBRARY")));
-    }
-
-    @Test
-    public void thatProjectInfoCreationFromMissingSbomThrowsNoException() {
-        assertThat(uploadBomMojo.createProjectInfo().isPresent(), is(equalTo(false)));
-    }
-
-    @Test
-    public void thatProjectInfoCreationFromOldSbomReturnsNoProjectInfo() {
-        uploadBomMojo.setBomLocation(UploadBomMojoTest.class.getResource("bom-1.1.xml").getFile());
-        assertThat(uploadBomMojo.createProjectInfo().isPresent(), is(equalTo(false)));
     }
 }
